requirements.txt
absl-py==1.0.0
asgiref==3.5.2
astunparse==1.6.3
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.1
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy
oauthlib==3.2.0
opencv-python
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-io-gcs-filesystem
termcolor
tf-estimator-nightly
tk
tomli
typing_extensions
urllib3
uvicorn
Werkzeug
wrapt
